requirements.txt
azure-core==1.33.0
azure-storage-blob==12.25.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
filelock==3.18.0
fsspec==2025.3.2
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
openai-whisper==20240930
pycparser==2.22
regex==2024.11.6
requests==2.32.3
six==1.17.0
sympy==1.13.3
tiktoken==0.9.0
torch==2.2.2
tqdm==4.67.1
typing_extensions==4.13.1
urllib3==2.3.0
